most of the way there
I was putting new things together and I am now mostly done.
