+++ Assets/TowerArt/Troops/Scenes/healer.tscn
@@ -1,6 +1,6 @@
 [gd_scene load_steps=17 format=3 uid="uid://dixum1rty8vlc"]
 
-[ext_resource type="Script" uid="uid://dbrxgcvdygqhc" path="res://Troops/Scripts/Healer.cs" id="1_v0y4t"]
+[ext_resource type="Script" uid="uid://dbrxgcvdygqhc" path="res://Assets/TowerArt/Troops/Scripts/Healer.cs" id="1_v0y4t"]
 [ext_resource type="Texture2D" uid="uid://cbo5h0wk0vyeh" path="res://Assets/CharacterArt/HealerCaveman.png" id="2_tamwb"]
 [ext_resource type="Texture2D" uid="uid://bfee6yv5v0ey3" path="res://Assets/CharacterArt/HealerMedieval.png" id="3_c6gsj"]
