+++ Assets/TowerArt/Troops/Scenes/ranged.tscn
@@ -1,0 +1,198 @@
+[gd_scene load_steps=23 format=3 uid="uid://bxye5em56ls6c"]
+
+[ext_resource type="Script" uid="uid://c4f10g1r2q511" path="res://Assets/TowerArt/Troops/Scripts/Ranged.cs" id="1_6my5x"]
+[ext_resource type="Texture2D" uid="uid://c6e37kt5ru71x" path="res://Assets/CharacterArt/RangedCavemanCharacter.png" id="2_w4ts2"]
+[ext_resource type="Texture2D" uid="uid://kh1mgeb7aixh" path="res://Assets/CharacterArt/MedievalRangedCharacter.png" id="3_hnljr"]
+[ext_resource type="Texture2D" uid="uid://bxee2iqm0fmt6" path="res://Assets/CharacterArt/RangedPresentCharacter.png" id="4_mgt3u"]
+[ext_resource type="Texture2D" uid="uid://bsw8gx45o4kn8" path="res://Assets/CharacterArt/RangedRobotCharacter.png" id="5_vu30a"]
+
+[sub_resource type="AtlasTexture" id="AtlasTexture_x3thf"]
+atlas = ExtResource("2_w4ts2")
+region = Rect2(0, 0, 16, 16)
+
+[sub_resource type="AtlasTexture" id="AtlasTexture_0c64o"]
+atlas = ExtResource("2_w4ts2")
+region = Rect2(16, 0, 16, 16)
+
+[sub_resource type="SpriteFrames" id="SpriteFrames_7mx6v"]
+animations = [{
+"frames": [{
+"duration": 1.0,
+"texture": SubResource("AtlasTexture_x3thf")
+}, {
+"duration": 1.0,
+"texture": SubResource("AtlasTexture_0c64o")
+}],
+"loop": true,
+"name": &"idle",
+"speed": 5.0
+}]
+
+[sub_resource type="AtlasTexture" id="AtlasTexture_kjh4o"]
+atlas = ExtResource("3_hnljr")
+region = Rect2(32, 0, 16, 16)
+
+[sub_resource type="AtlasTexture" id="AtlasTexture_wj5o0"]
+atlas = ExtResource("3_hnljr")
+region = Rect2(48, 0, 16, 16)
+
+[sub_resource type="SpriteFrames" id="SpriteFrames_om61b"]
+animations = [{
+"frames": [{
+"duration": 1.0,
+"texture": SubResource("AtlasTexture_kjh4o")
+}, {
+"duration": 1.0,
+"texture": SubResource("AtlasTexture_wj5o0")
+}],
+"loop": true,
+"name": &"idle",
+"speed": 5.0
+}]
+
+[sub_resource type="AtlasTexture" id="AtlasTexture_q5nrl"]
+atlas = ExtResource("4_mgt3u")
+region = Rect2(0, 0, 16, 16)
+
+[sub_resource type="AtlasTexture" id="AtlasTexture_2rje5"]
+atlas = ExtResource("4_mgt3u")
+region = Rect2(16, 0, 16, 16)
+
+[sub_resource type="AtlasTexture" id="AtlasTexture_pe4t5"]
+atlas = ExtResource("4_mgt3u")
+region = Rect2(32, 0, 16, 16)
+
+[sub_resource type="AtlasTexture" id="AtlasTexture_yycam"]
+atlas = ExtResource("4_mgt3u")
+region = Rect2(48, 0, 16, 16)
+
+[sub_resource type="SpriteFrames" id="SpriteFrames_mpgds"]
+animations = [{
+"frames": [{
+"duration": 1.0,
+"texture": SubResource("AtlasTexture_q5nrl")
+}, {
+"duration": 1.0,
+"texture": SubResource("AtlasTexture_2rje5")
+}, {
+"duration": 1.0,
+"texture": SubResource("AtlasTexture_pe4t5")
+}, {
+"duration": 1.0,
+"texture": SubResource("AtlasTexture_yycam")
+}],
+"loop": true,
+"name": &"idle",
+"speed": 5.0
+}]
+
+[sub_resource type="AtlasTexture" id="AtlasTexture_pvaa4"]
+atlas = ExtResource("5_vu30a")
+region = Rect2(16, 0, 16, 16)
+
+[sub_resource type="AtlasTexture" id="AtlasTexture_ehhsy"]
+atlas = ExtResource("5_vu30a")
+region = Rect2(32, 0, 16, 16)
+
+[sub_resource type="AtlasTexture" id="AtlasTexture_6g6sd"]
+atlas = ExtResource("5_vu30a")
+region = Rect2(48, 0, 16, 16)
+
+[sub_resource type="SpriteFrames" id="SpriteFrames_g05xj"]
+animations = [{
+"frames": [{
+"duration": 1.0,
+"texture": SubResource("AtlasTexture_pvaa4")
+}, {
+"duration": 1.0,
+"texture": SubResource("AtlasTexture_ehhsy")
+}, {
+"duration": 1.0,
+"texture": SubResource("AtlasTexture_6g6sd")
+}, {
+"duration": 1.0,
+"texture": SubResource("AtlasTexture_ehhsy")
+}],
+"loop": true,
+"name": &"idle",
+"speed": 5.0
+}]
+
+[sub_resource type="CircleShape2D" id="CircleShape2D_pcruc"]
+radius = 5.0
+
+[sub_resource type="CircleShape2D" id="CircleShape2D_y57lb"]
+radius = 5.0
+
+[node name="ranged" type="CharacterBody2D" groups=["Player"]]
+collision_mask = 0
+script = ExtResource("1_6my5x")
+
+[node name="Sprites" type="Node2D" parent="."]
+
+[node name="1" type="AnimatedSprite2D" parent="Sprites"]
+visible = false
+position = Vector2(-4.76837e-07, -4.76837e-07)
+sprite_frames = SubResource("SpriteFrames_7mx6v")
+animation = &"idle"
+autoplay = "idle"
+frame_progress = 0.539544
+
+[node name="2" type="AnimatedSprite2D" parent="Sprites"]
+visible = false
+position = Vector2(-4.76837e-07, -4.76837e-07)
+sprite_frames = SubResource("SpriteFrames_om61b")
+animation = &"idle"
+autoplay = "idle"
+frame_progress = 0.212046
+
+[node name="3" type="AnimatedSprite2D" parent="Sprites"]
+visible = false
+position = Vector2(-4.76837e-07, -4.76837e-07)
+sprite_frames = SubResource("SpriteFrames_mpgds")
+animation = &"idle"
+autoplay = "idle"
+frame_progress = 0.875325
+
+[node name="4" type="AnimatedSprite2D" parent="Sprites"]
+visible = false
+position = Vector2(-4.76837e-07, -4.76837e-07)
+sprite_frames = SubResource("SpriteFrames_g05xj")
+animation = &"idle"
+autoplay = "idle"
+frame_progress = 0.395135
+
+[node name="NavAgent" type="NavigationAgent2D" parent="."]
+
+[node name="Timer" type="Timer" parent="."]
+wait_time = 0.1
+autostart = true
+
+[node name="CollisionShape2D" type="CollisionShape2D" parent="."]
+visible = false
+shape = SubResource("CircleShape2D_pcruc")
+disabled = true
+
+[node name="Player1" type="Area2D" parent="."]
+visible = false
+collision_layer = 2
+collision_mask = 0
+
+[node name="CollisionShape2D" type="CollisionShape2D" parent="Player1"]
+shape = SubResource("CircleShape2D_y57lb")
+
+[node name="Player2" type="Area2D" parent="."]
+visible = false
+collision_layer = 4
+collision_mask = 0
+
+[node name="CollisionShape2D" type="CollisionShape2D" parent="Player2"]
+shape = SubResource("CircleShape2D_y57lb")
+
+[node name="Cooldown" type="Timer" parent="."]
+wait_time = 1.55
+one_shot = true
+
+[connection signal="navigation_finished" from="NavAgent" to="." method="on_path_finished"]
+[connection signal="timeout" from="Timer" to="." method="recalculate"]
+[connection signal="timeout" from="Cooldown" to="." method="on_cooldown"]
