+++ Assets/TowerArt/Troops/Scripts/Healer.cs
@@ -35,15 +35,37 @@ public partial class Healer : BaseTroop
 		foreach (AnimatedSprite2D child in GetNode<Node2D>("Sprites").GetChildren()){
 			child.Visible=false;
 		}
-		//GetNode<AnimatedSprite2D>("Sprites/1").Visible=true;
 		if(player1){
 			if(Player1Manager.upgradeLevel==1){
 				sprite = GetNode<AnimatedSprite2D>("Sprites/2");
 			 	GetNode<AnimatedSprite2D>("Sprites/2").Visible=true;
 			}
+			if(Player1Manager.upgradeLevel==2){
+				sprite = GetNode<AnimatedSprite2D>("Sprites/3");
+			 	GetNode<AnimatedSprite2D>("Sprites/3").Visible=true;
+			}
+			if(Player1Manager.upgradeLevel==3){
+				sprite = GetNode<AnimatedSprite2D>("Sprites/4");
+			 	GetNode<AnimatedSprite2D>("Sprites/4").Visible=true;
+			}
+		}
+		if(!player1){
+			if(Player2Manager.upgradeLevel==1){
+				sprite = GetNode<AnimatedSprite2D>("Sprites/2");
+			 	GetNode<AnimatedSprite2D>("Sprites/2").Visible=true;
+			}
+			if(Player2Manager.upgradeLevel==2){
+				sprite = GetNode<AnimatedSprite2D>("Sprites/3");
+			 	GetNode<AnimatedSprite2D>("Sprites/3").Visible=true;
+			}
+			if(Player2Manager.upgradeLevel==3){
+				sprite = GetNode<AnimatedSprite2D>("Sprites/4");
+			 	GetNode<AnimatedSprite2D>("Sprites/4").Visible=true;
+			}
 		}
 		
 		sprite.Visible=true;
+		sprite.Scale = new Vector2(-1, 1);
 	}
