+++ Assets/TowerArt/Troops/Scripts/Melee.cs
@@ -58,6 +58,20 @@ public partial class Melee : BaseTroop
 			 	GetNode<AnimatedSprite2D>("Sprites/4").Visible=true;
 			}
 		}
+		if(!player1){
+			if(Player2Manager.upgradeLevel==1){
+				sprite = GetNode<AnimatedSprite2D>("Sprites/2");
+			 	GetNode<AnimatedSprite2D>("Sprites/2").Visible=true;
+			}
+			if(Player2Manager.upgradeLevel==2){
+				sprite = GetNode<AnimatedSprite2D>("Sprites/3");
+			 	GetNode<AnimatedSprite2D>("Sprites/3").Visible=true;
+			}
+			if(Player2Manager.upgradeLevel==3){
+				sprite = GetNode<AnimatedSprite2D>("Sprites/4");
+			 	GetNode<AnimatedSprite2D>("Sprites/4").Visible=true;
+			}
+		}
 		sprite.Visible=true;
 		sprite.Scale = new Vector2(-1, 1);
 	}
