+++ Assets/TowerArt/Troops/Scripts/Ranged.cs
@@ -27,12 +27,42 @@ public partial class Ranged : BaseTroop
 		troopType=GameManager.Towers.Ranged;
 		health = maxHealth;
 		navAgent = GetNode<NavigationAgent2D>("NavAgent");
-		sprite = GetNode<AnimatedSprite2D>("Sprite");
 		cooldown = GetNode<Timer>("Cooldown"); updateHitboxes();
 		navAgent.TargetDesiredDistance = 300;
+		sprite = GetNode<AnimatedSprite2D>("Sprites/1");
 		await ToSignal(GetTree(), SceneTree.SignalName.ProcessFrame);
 		initialize();
 		fetchUpgrades();
 		TargetSet();
+		if(player1){
+			if(Player1Manager.upgradeLevel==1){
+				sprite = GetNode<AnimatedSprite2D>("Sprites/2");
+			 	GetNode<AnimatedSprite2D>("Sprites/2").Visible=true;
+			}
+			if(Player1Manager.upgradeLevel==2){
+				sprite = GetNode<AnimatedSprite2D>("Sprites/3");
+			 	GetNode<AnimatedSprite2D>("Sprites/3").Visible=true;
+			}
+			if(Player1Manager.upgradeLevel==3){
+				sprite = GetNode<AnimatedSprite2D>("Sprites/4");
+			 	GetNode<AnimatedSprite2D>("Sprites/4").Visible=true;
+			}
+		}
+		if(!player1){
+			if(Player2Manager.upgradeLevel==1){
+				sprite = GetNode<AnimatedSprite2D>("Sprites/2");
+			 	GetNode<AnimatedSprite2D>("Sprites/2").Visible=true;
+			}
+			if(Player2Manager.upgradeLevel==2){
+				sprite = GetNode<AnimatedSprite2D>("Sprites/3");
+			 	GetNode<AnimatedSprite2D>("Sprites/3").Visible=true;
+			}
+			if(Player2Manager.upgradeLevel==3){
+				sprite = GetNode<AnimatedSprite2D>("Sprites/4");
+			 	GetNode<AnimatedSprite2D>("Sprites/4").Visible=true;
+			}
+		}
+		sprite.Visible=true;
+		sprite.Scale = new Vector2(-1, 1);
 	}
 }
