+++ Scenes/base.tscn
@@ -3,10 +3,10 @@
 [ext_resource type="Texture2D" uid="uid://crdn1d2ugpbq8" path="res://Assets/TowerArt/player2Base.png" id="1_du40w"]
 [ext_resource type="Texture2D" uid="uid://cpr3ofkyasbcb" path="res://Assets/TowerArt/player1Base.png" id="1_qutmk"]
 [ext_resource type="Script" uid="uid://g0qp36p0meoh" path="res://Scripts/Base.cs" id="2_qutmk"]
-[ext_resource type="PackedScene" uid="uid://bjjy84kigbl4n" path="res://Troops/Scenes/melee.tscn" id="3_3v5je"]
-[ext_resource type="PackedScene" uid="uid://uvioh4v0m0i8" path="res://Troops/Scenes/brute.tscn" id="4_omojh"]
-[ext_resource type="PackedScene" uid="uid://bxye5em56ls6c" path="res://Troops/Scenes/ranged.tscn" id="5_yoopj"]
-[ext_resource type="PackedScene" uid="uid://dixum1rty8vlc" path="res://Troops/Scenes/healer.tscn" id="6_u35dc"]
+[ext_resource type="PackedScene" uid="uid://bjjy84kigbl4n" path="res://Assets/TowerArt/Troops/Scenes/melee.tscn" id="3_3v5je"]
+[ext_resource type="PackedScene" uid="uid://uvioh4v0m0i8" path="res://Assets/TowerArt/Troops/Scenes/brute.tscn" id="4_omojh"]
+[ext_resource type="PackedScene" uid="uid://bxye5em56ls6c" path="res://Assets/TowerArt/Troops/Scenes/ranged.tscn" id="5_yoopj"]
+[ext_resource type="PackedScene" uid="uid://dixum1rty8vlc" path="res://Assets/TowerArt/Troops/Scenes/healer.tscn" id="6_u35dc"]
 [ext_resource type="Texture2D" uid="uid://xihe12b6vlvk" path="res://Assets/Release.png" id="7_jlfiv"]
 [ext_resource type="Script" uid="uid://cws14ml62q3kc" path="res://Scripts/Controller.cs" id="8_kgqh3"]
 [ext_resource type="Texture2D" uid="uid://cl0xdsbm3g5ta" path="res://Assets/TowerArt/player2Smoke.png" id="10_du40w"]
