+++ miner.tscn
@@ -1,8 +1,8 @@
-[gd_scene load_steps=6 format=3 uid="uid://ssdf7h45x4d0"]
+[gd_scene load_steps=10 format=3 uid="uid://ssdf7h45x4d0"]
 
-[ext_resource type="Texture2D" uid="uid://bkqv5mmp0ataj" path="res://images/icon.jpg" id="1_u7vtn"]
 [ext_resource type="Script" uid="uid://cws14ml62q3kc" path="res://Scripts/Controller.cs" id="2_383ku"]
 [ext_resource type="Script" uid="uid://s627dlo2xofq" path="res://Miner.cs" id="2_u7vtn"]
+[ext_resource type="Texture2D" uid="uid://cbbrvijigdfl1" path="res://Assets/village.png" id="3_u7vtn"]
 
 [sub_resource type="CircleShape2D" id="CircleShape2D_nnceo"]
 radius = 578.947
@@ -10,12 +10,51 @@ radius = 578.947
 [sub_resource type="CircleShape2D" id="CircleShape2D_s42jq"]
 radius = 584.57
 
+[sub_resource type="AtlasTexture" id="AtlasTexture_383ku"]
+atlas = ExtResource("3_u7vtn")
+region = Rect2(48, 0, 48, 48)
+
+[sub_resource type="AtlasTexture" id="AtlasTexture_xecmj"]
+atlas = ExtResource("3_u7vtn")
+region = Rect2(0, 0, 48, 48)
+
+[sub_resource type="AtlasTexture" id="AtlasTexture_nnceo"]
+atlas = ExtResource("3_u7vtn")
+region = Rect2(96, 0, 48, 48)
+
+[sub_resource type="SpriteFrames" id="SpriteFrames_s42jq"]
+animations = [{
+"frames": [{
+"duration": 1.0,
+"texture": SubResource("AtlasTexture_383ku")
+}],
+"loop": true,
+"name": &"blue",
+"speed": 5.0
+}, {
+"frames": [{
+"duration": 1.0,
+"texture": SubResource("AtlasTexture_xecmj")
+}],
+"loop": true,
+"name": &"gray",
+"speed": 5.0
+}, {
+"frames": [{
+"duration": 1.0,
+"texture": SubResource("AtlasTexture_nnceo")
+}],
+"loop": true,
+"name": &"red",
+"speed": 5.0
+}]
+
 [node name="Miner" type="Sprite2D"]
-scale = Vector2(0.095, 0.095)
-texture = ExtResource("1_u7vtn")
 script = ExtResource("2_u7vtn")
 
 [node name="Player-1" type="Area2D" parent="."]
+visible = false
+scale = Vector2(0.115, 0.115)
 collision_layer = 0
 collision_mask = 0
 
@@ -23,6 +62,8 @@ collision_mask = 0
 shape = SubResource("CircleShape2D_nnceo")
 
 [node name="Player-2" type="Area2D" parent="."]
+position = Vector2(-1, -11)
+scale = Vector2(0.115, 0.115)
 collision_layer = 0
 collision_mask = 0
 
@@ -30,6 +71,8 @@ collision_mask = 0
 shape = SubResource("CircleShape2D_s42jq")
 
 [node name="Player-None" type="Area2D" parent="."]
+position = Vector2(-1, -11)
+scale = Vector2(0.115, 0.115)
 collision_layer = 2048
 collision_mask = 0
 
@@ -38,11 +81,11 @@ position = Vector2(0, -8.25806)
 shape = SubResource("CircleShape2D_s42jq")
 
 [node name="Button" type="Button" parent="."]
-offset_left = -288.0
-offset_top = -289.032
-offset_right = 284.0
-offset_bottom = 282.968
-scale = Vector2(0.895, 0.895)
+offset_left = -277.0
+offset_top = -293.0
+offset_right = 295.0
+offset_bottom = 279.0
+scale = Vector2(0.075, 0.075)
 pivot_offset = Vector2(273.077, 280.086)
 mouse_filter = 2
 flat = true
@@ -58,15 +101,27 @@ scale = Vector2(11.5268, 24.6106)
 polygon = PackedVector2Array(45.146, 23.1584, 45.146, -0.0863104, -4.48339, -0.0863104, -4.48339, 23.1584)
 
 [node name="Player2" type="Area2D" parent="Button"]
+visible = false
 collision_layer = 512
 collision_mask = 0
 
 [node name="CollisionPolygon2D" type="CollisionPolygon2D" parent="Button/Player2"]
 polygon = PackedVector2Array(572.073, 571.874, 572.073, -0.193542, 0.00622559, -0.193542, 0.00622559, 571.874)
 
-[node name="Sprite2D" type="Sprite2D" parent="Button"]
-position = Vector2(286.04, 276.613)
-texture = ExtResource("1_u7vtn")
+[node name="Control" type="Control" parent="Button"]
+clip_contents = true
+anchors_preset = 0
+offset_left = 5.0
+offset_top = -1.0
+offset_right = 579.0
+offset_bottom = 565.0
+
+[node name="Sprite2D" type="AnimatedSprite2D" parent="Button/Control"]
+position = Vector2(276.102, 292.548)
+scale = Vector2(11.9333, 11.9333)
+sprite_frames = SubResource("SpriteFrames_s42jq")
+animation = &"gray"
+autoplay = "gray"
 offset = Vector2(-3.40979e-05, 0)
 
 [node name="Timer" type="Timer" parent="."]
